+++ domain-models/src/main/java/com/jeanloth/project/android/kotlin/domain_models/entities/AnalysisList.kt
@@ -3,5 +3,5 @@ package com.jeanloth.project.android.kotlin.domain_models.entities
 data class AnalysisList(
     val label: String,
     val number: Int = 0,
-    val totalAmount: Double = 0.0,
+    val totalAmount: Int = 0,
 )
